Remove scripts directory in favor of nodes subpackage
Since these modules will no longer be called as scripts in ROS 2,
move them to a nodes subpackage, and remove the option to call them
as executables.
