+++ setup.py
@@ -8,12 +8,11 @@ SHARE_DIR = os.path.join("share", PACKAGE_NAME)
 setup(
     name=PACKAGE_NAME,
     version='0.5.1',
-    packages=["scripts", "libnmea_navsat_driver"],
+    packages=["libnmea_navsat_driver", "libnmea_navsat_driver.nodes"],
     data_files=[
         (os.path.join(SHARE_DIR, "launch"), glob(os.path.join("launch", "*.launch.py"))),
         (os.path.join(SHARE_DIR, "config"), glob(os.path.join("config", "*.yaml")))],
-    package_dir={'': 'src',
-                 "scripts": "scripts"},
+    package_dir={'': 'src',},
     py_modules=[],
     zip_safe=True,
     install_requires=['setuptools',
@@ -27,9 +26,9 @@ setup(
     description='Package to parse NMEA strings and publish a very simple GPS message.',
     license='BSD',
     entry_points={
-        'console_scripts': ['nmea_serial_driver = scripts.nmea_serial_driver:main',
-                            'nmea_socket_driver = scripts.nmea_socket_driver:main',
-                            'nmea_topic_driver = scripts.nmea_topic_driver:main',
-                            'nmea_topic_serial_reader = scripts.nmea_topic_serial_reader:main'],
+        'console_scripts': ['nmea_serial_driver = libnmea_navsat_driver.nodes.nmea_serial_driver:main',
+                            'nmea_socket_driver = libnmea_navsat_driver.nodes.nmea_socket_driver:main',
+                            'nmea_topic_driver = libnmea_navsat_driver.nodes.nmea_topic_driver:main',
+                            'nmea_topic_serial_reader = libnmea_navsat_driver.nodes.nmea_topic_serial_reader:main'],
     }
 )
