+++ src/libnmea_navsat_driver/nodes/nmea_topic_driver.py
@@ -1,5 +1,3 @@
-#! /usr/bin/env python
-
 # Software License Agreement (BSD License)
 #
 # Copyright (c) 2013, Eric Perko
@@ -32,43 +30,33 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-import serial
 
+from nmea_msgs.msg import Sentence
 import rclpy
 
 from libnmea_navsat_driver.driver import Ros2NMEADriver
+
+
+def nmea_sentence_callback(nmea_sentence, driver):
+    try:
+        driver.add_sentence(nmea_sentence.sentence, frame_id=nmea_sentence.header.frame_id,
+                            timestamp=nmea_sentence.header.stamp)
+    except ValueError as e:
+        rclpy.get_logger().warn(
+            "Value error, likely due to missing fields in the NMEA message. Error was: %s. "
+            "Please report this issue at github.com/ros-drivers/nmea_navsat_driver, including a bag file with "
+            "the NMEA sentences that caused it." % e)
 
 
 def main(args=None):
     rclpy.init(args=args)
 
     driver = Ros2NMEADriver()
-    frame_id = driver.get_frame_id()
+    driver.get_frame_id()
 
-    serial_port = driver.get_parameter('port').value or '/dev/ttyUSB0'
-    serial_baud = driver.get_parameter('baud').value or 4800
+    driver.create_subscription(
+        Sentence, 'nmea_sentence', nmea_sentence_callback, driver)
 
-    try:
-        GPS = serial.Serial(port=serial_port, baudrate=serial_baud, timeout=2)
-        try:
-            while rclpy.ok():
-                data = GPS.readline().strip()
-                try:
-                    if isinstance(data, bytes):
-                        data = data.decode("utf-8")
-                    driver.add_sentence(data, frame_id)
-                except ValueError as e:
-                    driver.get_logger().warn(
-                        "Value error, likely due to missing fields in the NMEA message. Error was: %s. "
-                        "Please report this issue at github.com/ros-drivers/nmea_navsat_driver, including a bag file "
-                        "with the NMEA sentences that caused it." % e)
+    rclpy.spin(node)
 
-        except Exception as e:
-            driver.get_logger().error("Ros error: {0}".format(e))
-            GPS.close()  # Close GPS serial port
-    except serial.SerialException as ex:
-        driver.get_logger().fatal("Could not open serial port: I/O error({0}): {1}".format(ex.errno, ex.strerror))
-
-
-if __name__ == '__main__':
-    main()
+    rclpy.shutdown()
