+++ src/libnmea_navsat_driver/nodes/nmea_topic_serial_reader.py
@@ -1,5 +1,3 @@
-#! /usr/bin/env python
-
 # Software License Agreement (BSD License)
 #
 # Copyright (c) 2013, Eric Perko
@@ -32,37 +30,41 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-import rclpy
+import serial
 
 from nmea_msgs.msg import Sentence
+import rclpy
 
 from libnmea_navsat_driver.driver import Ros2NMEADriver
-
-
-def nmea_sentence_callback(nmea_sentence, driver):
-    try:
-        driver.add_sentence(nmea_sentence.sentence, frame_id=nmea_sentence.header.frame_id,
-                            timestamp=nmea_sentence.header.stamp)
-    except ValueError as e:
-        rclpy.get_logger().warn(
-            "Value error, likely due to missing fields in the NMEA message. Error was: %s. "
-            "Please report this issue at github.com/ros-drivers/nmea_navsat_driver, including a bag file with "
-            "the NMEA sentences that caused it." % e)
 
 
 def main(args=None):
     rclpy.init(args=args)
 
     driver = Ros2NMEADriver()
-    driver.get_frame_id()
 
-    driver.create_subscription(
-        Sentence, 'nmea_sentence', nmea_sentence_callback, driver)
+    nmea_pub = driver.create_publisher(Sentence, "nmea_sentence")
 
-    rclpy.spin(node)
+    serial_port = driver.get_parameter('port').value or '/dev/ttyUSB0'
+    serial_baud = driver.get_parameter('baud').value or 4800
 
-    rclpy.shutdown()
+    # Get the frame_id
+    frame_id = driver.get_frame_id()
 
+    try:
+        GPS = serial.Serial(port=serial_port, baudrate=serial_baud, timeout=2)
+        try:
+            while rclpy.ok():
+                data = GPS.readline().strip()
 
-if __name__ == '__main__':
-    main()
+                sentence = Sentence()
+                sentence.header.stamp = driver.get_clock().now().to_msg()
+                sentence.header.frame_id = frame_id
+                sentence.sentence = data
+                nmea_pub.publish(sentence)
+
+        except Exception as e:
+            driver.get_logger().error("Ros error: {0}".format(e))
+            GPS.close()  # Close GPS serial port
+    except serial.SerialException as ex:
+        driver.get_logger().fatal("Could not open serial port: I/O error({0}): {1}".format(ex.errno, ex.strerror))
